pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lucy"
version = "0.1.0"
description = "AI coworker for Slack — built on OpenClaw"
requires-python = ">=3.12"
dependencies = [
    "slack-bolt>=1.21,<2",
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pydantic-settings>=2.7,<3",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "psycopg2-binary>=2.9,<3",
    "qdrant-client>=1.12,<2",
    "mem0ai>=0.1,<1",
    "litellm>=1.59,<2",
    "structlog>=24.4,<25",
    "httpx>=0.28,<1",
    "python-dateutil>=2.9,<3",
    "croniter>=6.0,<7",
    "composio-core>=0.5,<1",
    "composio-openai>=0.5,<1",
    "psutil>=6.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
